idemix2: match MPI datatypes to the working precision WP
Buffers declared real(kind=WP) were communicated as MPI_DOUBLE_PRECISION:
the read_other_NetCDF/-_3d broadcasts (which also serve the pre-existing
runoff/SSS/chlorophyll/ice readers), the IDEMIX2 init/diagnostic
Allreduces (one of which normalizes the bottom tidal forcing), the fbin
halo-exchange dummies (declared real64) and the fbin MPI type creation.
In a USE_SINGLE_PRECISION build (WP=4) each such transfer reads/writes
twice the buffer - heap corruption on the broadcasts, garbage physics in
the forcing normalization. Use MPI_WP and real(kind=WP) throughout, as
main does since the single-precision support landed; double-precision
builds are unchanged. The fbin gather-to-master routines keep real64,
matching the convention of the neighbouring gather routines.

+++ src/gen_halo_exchange.F90
@@ -2976,7 +2976,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target :: partit
-real(real64),   intent(inout)         :: nod_array(:,:)
+real(kind=WP),   intent(inout)         :: nod_array(:,:)
 if (partit%npes > 1) then
    call exchange_nod2D_fbin_begin(nod_array, partit)
    call exchange_nod_end(partit)
@@ -2989,7 +2989,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target :: partit
-real(real64),   intent(inout)         :: nod_array(:,:)
+real(kind=WP),   intent(inout)         :: nod_array(:,:)
 integer                               :: n, sn, rn
 #include "associate_part_def.h"
 #include "associate_part_ass.h"
@@ -3026,7 +3026,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target :: partit
-real(real64),   intent(inout)         :: nod_array(:,:,:)
+real(kind=WP),   intent(inout)         :: nod_array(:,:,:)
 if (partit%npes > 1) then
    call exchange_nod3D_fbin_begin(nod_array, partit)
    call exchange_nod_end(partit)
@@ -3039,7 +3039,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target :: partit
-real(real64),   intent(inout)         :: nod_array(:,:,:)
+real(kind=WP),   intent(inout)         :: nod_array(:,:,:)
 integer                               :: n, sn, rn, n_val
 #include "associate_part_def.h"
 #include "associate_part_ass.h"
@@ -3083,7 +3083,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target  :: partit
-real(real64),   intent(inout)          :: elem_array(:,:)
+real(kind=WP),   intent(inout)          :: elem_array(:,:)
 #include "associate_part_def.h"
 #include "associate_part_ass.h"
 
@@ -3099,7 +3099,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target  :: partit
-real(real64),   intent(inout)          :: elem_array(:,:)
+real(kind=WP),   intent(inout)          :: elem_array(:,:)
 integer                                :: n, sn, rn
 #include "associate_part_def.h"
 #include "associate_part_ass.h"
@@ -3157,7 +3157,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target  :: partit
-real(real64),   intent(inout)          :: elem_array(:,:,:)
+real(kind=WP),   intent(inout)          :: elem_array(:,:,:)
 #include "associate_part_def.h"
 #include "associate_part_ass.h"
 
@@ -3173,7 +3173,7 @@ USE MOD_PARTIT
 USE MOD_PARSUP
 IMPLICIT NONE
 type(t_partit), intent(inout), target  :: partit
-real(real64),   intent(inout)          :: elem_array(:,:,:)
+real(kind=WP),   intent(inout)          :: elem_array(:,:,:)
 integer                                :: n, sn, rn, n_val
 #include "associate_part_def.h"
 #include "associate_part_ass.h"
